+++ components/media/EditMedia.vue
@@ -163,7 +163,24 @@
             <div>{{ $t(mediaTranslationPrefix + 'EditMedia.downloadMedia') }}</div>
           </div>
 
-          <div v-if="!(lockedStatus === 'locked' && media.author !== sectionsUserIdProp)">
+          <div v-if="media.type === 'document'">
+            <div class="flex w-350px h-200px justify-center items-center object-cover relative" :style="hiddenContainerStyle">
+              <div class="flex flex-col items-center gap-4">
+                <span class="icon-mediaDocument text-6xl"></span>
+              </div>
+              <div class="absolute top-1/3 left-1/3 -translate-x-1/3 -translate-y-1/3" @click="$refs.imagePick.click()">
+                <span class="icon-reload text-8xl cursor-pointer"></span>
+                <input
+                  ref="imagePick"
+                  type="file"
+                  class="hidden"
+                  accept=".pdf, .doc, .docx, .zip, .json, .css, .scss, .xlsx, .xlsb, .xltx"
+                  @change="onFileSelected"
+                />
+              </div>
+            </div>
+          </div>
+          <div v-else-if="!(lockedStatus === 'locked' && media.author !== sectionsUserIdProp)">
             <div v-if="media.files && media.files[0].url !== ''" class="relative w-max">
               <img :src="media.files[0].url" alt="" class="rounded-md w-400px">
               <div class="absolute top-1/3 left-1/3 -translate-x-1/3 -translate-y-1/3" @click="$refs.imagePick.click()">
@@ -172,6 +189,7 @@
                   ref="imagePick"
                   type="file"
                   class="hidden"
+                  accept="image/*"
                   @change="onFileSelected"
                 />
               </div>
@@ -274,6 +292,10 @@ export default {
     mediasPath: {
       type: String,
       default: ""
+    },
+    hiddenContainerStyle: {
+      type: String,
+      default: "background: #61035B"
     },
     mediaByIdResponseProp: {
       type: Object,
@@ -544,6 +566,7 @@ export default {
       const data = new FormData();
       if(this.file) {
         data.append('files[1][file]', this.file);
+        data.append('type', this.file.type && this.file.type.includes('image') ? 'image' : 'document');
       }
       if(this.media.title && this.media.title !== '') data.append('title', this.media.title);
       if(this.media.seo_tag && this.media.seo_tag !== '') data.append('seo_tag', this.media.seo_tag);
@@ -657,7 +680,7 @@ export default {
     backClicked() {
       this.isEditingMedia = false
       if (this.mediasPath) {
-        this.$router.push(this.localePath({path: this.mediasPath, query: {filters: this.$route.query.filters}}))
+        this.$router.push(this.localePath({path: this.mediasPath, query: {filters: this.$route.query.filters, folder: this.$route.query.folder}}))
       } else this.$emit('updateMediaComponent', {name: 'MediaListMedias'})
     },
     handleFileName(fileName) {