+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 - [x] Bug fixes for media meta components
 
+- [x] Add support for documents upload/document folders
+
 
 
 ### v1.0.2 (30/6/2023)